+++ src/db/BaseQuery.php
@@ -16,7 +16,6 @@ use think\Collection;
 use think\db\exception\DataNotFoundException;
 use think\db\exception\DbException as Exception;
 use think\db\exception\ModelNotFoundException;
-use think\db\exception\PDOException;
 use think\helper\Str;
 use think\Model;
 use think\Paginator;
@@ -24,7 +23,7 @@ use think\Paginator;
 /**
  * 数据查询基础类
  */
-class BaseQuery
+abstract class BaseQuery
 {
     use concern\TimeFieldQuery;
     use concern\AggregateQuery;
@@ -104,64 +103,13 @@ class BaseQuery
     }
 
     /**
-     * 利用__call方法实现一些特殊的Model方法
-     * @access public
-     * @param string $method 方法名称
-     * @param array  $args   调用参数
-     * @return mixed
-     * @throws Exception
-     */
-    public function __call(string $method, array $args)
-    {
-        if (strtolower(substr($method, 0, 5)) == 'getby') {
-            // 根据某个字段获取记录
-            $field = Str::snake(substr($method, 5));
-            return $this->where($field, '=', $args[0])->find();
-        } elseif (strtolower(substr($method, 0, 10)) == 'getfieldby') {
-            // 根据某个字段获取记录的某个值
-            $name = Str::snake(substr($method, 10));
-            return $this->where($name, '=', $args[0])->value($args[1]);
-        } elseif (strtolower(substr($method, 0, 7)) == 'whereor') {
-            $name = Str::snake(substr($method, 7));
-            array_unshift($args, $name);
-            return call_user_func_array([$this, 'whereOr'], $args);
-        } elseif (strtolower(substr($method, 0, 5)) == 'where') {
-            $name = Str::snake(substr($method, 5));
-            array_unshift($args, $name);
-            return call_user_func_array([$this, 'where'], $args);
-        } elseif ($this->model && method_exists($this->model, 'scope' . $method)) {
-            // 动态调用命名范围
-            $method = 'scope' . $method;
-            array_unshift($args, $this);
-
-            call_user_func_array([$this->model, $method], $args);
-            return $this;
-        } else {
-            throw new Exception('method not exist:' . static::class . '->' . $method);
-        }
-    }
-
-    /**
      * 获取当前的数据库Connection对象
      * @access public
-     * @return Connection
+     * @return ConnectionInterface
      */
     public function getConnection()
     {
         return $this->connection;
-    }
-
-    /**
-     * 设置当前的数据库Connection对象
-     * @access public
-     * @param Connection $connection 数据库连接对象
-     * @return $this
-     */
-    public function setConnection(Connection $connection)
-    {
-        $this->connection = $connection;
-
-        return $this;
     }
 
     /**
@@ -212,6 +160,18 @@ class BaseQuery
         $name = $name ?: $this->name;
 
         return $this->prefix . Str::snake($name);
+    }
+
+    /**
+     * 设置字段类型信息
+     * @access public
+     * @param array $type 字段类型信息
+     * @return $this
+     */
+    public function setFieldType(array $type)
+    {
+        $this->options['field_type'] = $type;
+        return $this;
     }
 
     /**
@@ -403,19 +363,6 @@ class BaseQuery
     }
 
     /**
-     * 表达式方式指定查询字段
-     * @access public
-     * @param string $field 字段名
-     * @return $this
-     */
-    public function fieldRaw(string $field)
-    {
-        $this->options['field'][] = new Raw($field);
-
-        return $this;
-    }
-
-    /**
      * 设置数据
      * @access public
      * @param array $data 数据
@@ -425,46 +372,6 @@ class BaseQuery
     {
         $this->options['data'] = $data;
 
-        return $this;
-    }
-
-    /**
-     * 字段值增长
-     * @access public
-     * @param string  $field    字段名
-     * @param float   $step     增长值
-     * @return $this
-     */
-    public function inc(string $field, float $step = 1)
-    {
-        $this->options['data'][$field] = ['INC', $step];
-
-        return $this;
-    }
-
-    /**
-     * 字段值减少
-     * @access public
-     * @param string  $field    字段名
-     * @param float   $step     增长值
-     * @return $this
-     */
-    public function dec(string $field, float $step = 1)
-    {
-        $this->options['data'][$field] = ['DEC', $step];
-        return $this;
-    }
-
-    /**
-     * 使用表达式设置数据
-     * @access public
-     * @param string $field 字段名
-     * @param string $value 字段值
-     * @return $this
-     */
-    public function exp(string $field, string $value)
-    {
-        $this->options['data'][$field] = new Raw($value);
         return $this;
     }
 
@@ -510,6 +417,107 @@ class BaseQuery
     public function page(int $page, int $listRows = null)
     {
         $this->options['page'] = [$page, $listRows];
+
+        return $this;
+    }
+
+    /**
+     * 指定当前操作的数据表
+     * @access public
+     * @param mixed $table 表名
+     * @return $this
+     */
+    public function table($table)
+    {
+        if (is_string($table)) {
+            if (strpos($table, ')')) {
+                // 子查询
+            } elseif (false === strpos($table, ',')) {
+                if (strpos($table, ' ')) {
+                    list($item, $alias) = explode(' ', $table);
+                    $table              = [];
+                    $this->alias([$item => $alias]);
+                    $table[$item] = $alias;
+                }
+            } else {
+                $tables = explode(',', $table);
+                $table  = [];
+
+                foreach ($tables as $item) {
+                    $item = trim($item);
+                    if (strpos($item, ' ')) {
+                        list($item, $alias) = explode(' ', $item);
+                        $this->alias([$item => $alias]);
+                        $table[$item] = $alias;
+                    } else {
+                        $table[] = $item;
+                    }
+                }
+            }
+        } elseif (is_array($table)) {
+            $tables = $table;
+            $table  = [];
+
+            foreach ($tables as $key => $val) {
+                if (is_numeric($key)) {
+                    $table[] = $val;
+                } else {
+                    $this->alias([$key => $val]);
+                    $table[$key] = $val;
+                }
+            }
+        }
+
+        $this->options['table'] = $table;
+
+        return $this;
+    }
+
+    /**
+     * 指定排序 order('id','desc') 或者 order(['id'=>'desc','create_time'=>'desc'])
+     * @access public
+     * @param string|array|Raw $field 排序字段
+     * @param string           $order 排序
+     * @return $this
+     */
+    public function order($field, string $order = '')
+    {
+        if (empty($field)) {
+            return $this;
+        } elseif ($field instanceof Raw) {
+            $this->options['order'][] = $field;
+            return $this;
+        }
+
+        if (is_string($field)) {
+            if (!empty($this->options['via'])) {
+                $field = $this->options['via'] . '.' . $field;
+            }
+            if (strpos($field, ',')) {
+                $field = array_map('trim', explode(',', $field));
+            } else {
+                $field = empty($order) ? $field : [$field => $order];
+            }
+        } elseif (!empty($this->options['via'])) {
+            foreach ($field as $key => $val) {
+                if (is_numeric($key)) {
+                    $field[$key] = $this->options['via'] . '.' . $val;
+                } else {
+                    $field[$this->options['via'] . '.' . $key] = $val;
+                    unset($field[$key]);
+                }
+            }
+        }
+
+        if (!isset($this->options['order'])) {
+            $this->options['order'] = [];
+        }
+
+        if (is_array($field)) {
+            $this->options['order'] = array_merge($this->options['order'], $field);
+        } else {
+            $this->options['order'][] = $field;
+        }
 
         return $this;
     }
@@ -679,168 +687,6 @@ class BaseQuery
             'data'   => $result,
             'lastId' => $last[$key],
         ];
-    }
-
-    /**
-     * 表达式方式指定当前操作的数据表
-     * @access public
-     * @param mixed $table 表名
-     * @return $this
-     */
-    public function tableRaw(string $table)
-    {
-        $this->options['table'] = new Raw($table);
-
-        return $this;
-    }
-
-    /**
-     * 指定当前操作的数据表
-     * @access public
-     * @param mixed $table 表名
-     * @return $this
-     */
-    public function table($table)
-    {
-        if (is_string($table)) {
-            if (strpos($table, ')')) {
-                // 子查询
-            } elseif (false === strpos($table, ',')) {
-                if (strpos($table, ' ')) {
-                    list($item, $alias) = explode(' ', $table);
-                    $table              = [];
-                    $this->alias([$item => $alias]);
-                    $table[$item] = $alias;
-                }
-            } else {
-                $tables = explode(',', $table);
-                $table  = [];
-
-                foreach ($tables as $item) {
-                    $item = trim($item);
-                    if (strpos($item, ' ')) {
-                        list($item, $alias) = explode(' ', $item);
-                        $this->alias([$item => $alias]);
-                        $table[$item] = $alias;
-                    } else {
-                        $table[] = $item;
-                    }
-                }
-            }
-        } elseif (is_array($table)) {
-            $tables = $table;
-            $table  = [];
-
-            foreach ($tables as $key => $val) {
-                if (is_numeric($key)) {
-                    $table[] = $val;
-                } else {
-                    $this->alias([$key => $val]);
-                    $table[$key] = $val;
-                }
-            }
-        }
-
-        $this->options['table'] = $table;
-
-        return $this;
-    }
-
-    /**
-     * 指定排序 order('id','desc') 或者 order(['id'=>'desc','create_time'=>'desc'])
-     * @access public
-     * @param string|array|Raw $field 排序字段
-     * @param string           $order 排序
-     * @return $this
-     */
-    public function order($field, string $order = '')
-    {
-        if (empty($field)) {
-            return $this;
-        } elseif ($field instanceof Raw) {
-            $this->options['order'][] = $field;
-            return $this;
-        }
-
-        if (is_string($field)) {
-            if (!empty($this->options['via'])) {
-                $field = $this->options['via'] . '.' . $field;
-            }
-            if (strpos($field, ',')) {
-                $field = array_map('trim', explode(',', $field));
-            } else {
-                $field = empty($order) ? $field : [$field => $order];
-            }
-        } elseif (!empty($this->options['via'])) {
-            foreach ($field as $key => $val) {
-                if (is_numeric($key)) {
-                    $field[$key] = $this->options['via'] . '.' . $val;
-                } else {
-                    $field[$this->options['via'] . '.' . $key] = $val;
-                    unset($field[$key]);
-                }
-            }
-        }
-
-        if (!isset($this->options['order'])) {
-            $this->options['order'] = [];
-        }
-
-        if (is_array($field)) {
-            $this->options['order'] = array_merge($this->options['order'], $field);
-        } else {
-            $this->options['order'][] = $field;
-        }
-
-        return $this;
-    }
-
-    /**
-     * 表达式方式指定Field排序
-     * @access public
-     * @param string $field 排序字段
-     * @param array  $bind  参数绑定
-     * @return $this
-     */
-    public function orderRaw(string $field, array $bind = [])
-    {
-        if (!empty($bind)) {
-            $this->bindParams($field, $bind);
-        }
-
-        $this->options['order'][] = new Raw($field);
-
-        return $this;
-    }
-
-    /**
-     * 指定Field排序 orderField('id',[1,2,3],'desc')
-     * @access public
-     * @param string $field  排序字段
-     * @param array  $values 排序值
-     * @param string $order  排序 desc/asc
-     * @return $this
-     */
-    public function orderField(string $field, array $values, string $order = '')
-    {
-        if (!empty($values)) {
-            $values['sort'] = $order;
-
-            $this->options['order'][$field] = $values;
-        }
-
-        return $this;
-    }
-
-    /**
-     * 随机排序
-     * @access public
-     * @return $this
-     */
-    public function orderRand()
-    {
-        $this->options['order'][] = '[rand]';
-        return $this;
     }
 
     /**
@@ -1083,7 +929,6 @@ class BaseQuery
      * @param array  $fields 要插入的数据表字段名
      * @param string $table  要插入的数据表名
      * @return integer
-     * @throws PDOException
      */
     public function selectInsert(array $fields, string $table): int
     {
@@ -1096,7 +941,6 @@ class BaseQuery
      * @param mixed $data 数据
      * @return integer
      * @throws Exception
-     * @throws PDOException
      */
     public function update(array $data = []): int
     {
@@ -1126,7 +970,6 @@ class BaseQuery
      * @param mixed $data 表达式 true 表示强制删除
      * @return int
      * @throws Exception
-     * @throws PDOException
      */
     public function delete($data = null): int
     {
@@ -1229,65 +1072,6 @@ class BaseQuery
         }
 
         return $result;
-    }
-
-    /**
-     * 分批数据返回处理
-     * @access public
-     * @param integer      $count    每次处理的数据数量
-     * @param callable     $callback 处理回调方法
-     * @param string|array $column   分批处理的字段名
-     * @param string       $order    字段排序
-     * @return bool
-     * @throws Exception
-     */
-    public function chunk(int $count, callable $callback, $column = null, string $order = 'asc'): bool
-    {
-        $options = $this->getOptions();
-        $column  = $column ?: $this->getPk();
-
-        if (isset($options['order'])) {
-            unset($options['order']);
-        }
-
-        $bind = $this->bind;
-
-        if (is_array($column)) {
-            $times = 1;
-            $query = $this->options($options)->page($times, $count);
-        } else {
-            $query = $this->options($options)->limit($count);
-
-            if (strpos($column, '.')) {
-                list($alias, $key) = explode('.', $column);
-            } else {
-                $key = $column;
-            }
-        }
-
-        $resultSet = $query->order($column, $order)->select();
-
-        while (count($resultSet) > 0) {
-            if (false === call_user_func($callback, $resultSet)) {
-                return false;
-            }
-
-            if (isset($times)) {
-                $times++;
-                $query = $this->options($options)->page($times, $count);
-            } else {
-                $end    = $resultSet->pop();
-                $lastId = is_array($end) ? $end[$key] : $end->getData($key);
-
-                $query = $this->options($options)
-                    ->limit($count)
-                    ->where($column, 'asc' == strtolower($order) ? '>' : '<', $lastId);
-            }
-
-            $resultSet = $query->bind($bind)->order($column, $order)->select();
-        }
-
-        return true;
     }
 
     /**
